Fixed sonar.

+++ src/main/java/g2101_2200/s2195_append_k_integers_with_minimal_sum/Solution.java
@@ -18,7 +18,7 @@ public class Solution {
                 n++;
             }
         }
-        long t = n + k;
+        long t = n + (long) k;
         return (1 + t) * t / 2 - sum;
     }
 }
